Create climbing the leaderboard.cpp
Alice is playing an arcade game and wants to climb to the top of the leader-board and wants to track her ranking. The game uses Dense Ranking, so its leader-board works like this:

The player with the highest score is ranked number 1 on the leader board.
Players who have equal scores receive the same ranking number, and the next player(s) receive the immediately following ranking number.