Create Append and Delete.cpp
https://www.hackerrank.com/challenges/append-and-delete/problem

Explanation:
we have to check few cases............
case1: Minimum number of Operation required is greater than k. In this case it is not possible to convert s to t;
case2: If Mop is less than k then this case is checked. Now it is possible to convert s to t if both MOP and k is either even or both are odd.
case3: If case2 is not satisfied then we can get rid of the even-odd problem if we can make 's' empty by deleting all element. Now in the problem it is given that *performing delete operation on an empty string results in an empty string*, so we can avoid the even-odd problem by performing deleting operation as required. In this case MOP = s.lenght() + t.length() + 1;